requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.7
fastapi-jwt-auth==0.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
MarkupSafe==3.0.2
psycopg==3.2.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
